Move testSingle() to new class for test
